changed long to int
git-svn-id: file:///home/SUBVERSION/jas/trunk@1581 eafcb37d-f50b-0410-b8f7-d4cb9effa109

+++ src/edu/jas/structure/ProductRing.java
@@ -150,7 +150,7 @@ public class ProductRing<C extends RingElem<C> >
     /** Get the number of factors of this ring.
      * @return nCopies or ringList.size().
      */
-    public long length() {
+    public int length() {
         if ( nCopies != 0 ) {
            return nCopies;
         } else {
